Sherlock_Times/requirements.txt
streamlit>=1.35.0
feedparser>=6.0.11
streamlit-autorefresh>=0.0.1
newspaper3k>=0.2.8
lxml[html_clean]>=5.3.0
lxml_html_clean>=0.1.0
beautifulsoup4>=4.12.0
vaderSentiment>=3.3.2